Refactor: Initialize Vite, integrate Supabase Auth, and prep for Cloudflare
Key changes:
- Restructured the project from a single HTML file with CDN React to a standard Vite + React setup.
- Removed `<!DOCTYPE html>.js` and `electroshop.js` (raw React), using `src/App.jsx` as the main component.
- Integrated Supabase for your authentication:
    - Added Supabase client and configured it to use environment variables (`VITE_SUPABASE_URL`, `VITE_SUPABASE_ANON_KEY`).
    - Implemented login/logout functionality using Supabase Auth UI.
    - Application now displays your session information.
- Prepared for Cloudflare Pages deployment:
    - Updated `README.md` with detailed instructions for local development, build, and Cloudflare Pages deployment, including environment variable setup and Supabase dashboard configuration.
- Ensured Tailwind CSS is available for styling.

The application now has a modern build process, robust authentication, and is ready for deployment.
