+++ README.md
@@ -1,0 +1,60 @@
+# ElectroShop - React E-commerce Frontend
+
+This is a simple e-commerce frontend application built with React and Vite, featuring authentication using Supabase.
+
+## Development
+
+1.  **Clone the repository:**
+    ```bash
+    git clone <your-repo-url>
+    cd <repo-name>
+    ```
+
+2.  **Install dependencies:**
+    ```bash
+    npm install
+    ```
+
+3.  **Set up environment variables (for local development):**
+    Create a `.env` file in the root of the project and add your Supabase credentials:
+    ```
+    VITE_SUPABASE_URL=https://your-project-ref.supabase.co
+    VITE_SUPABASE_ANON_KEY=your-supabase-anon-key
+    ```
+    Replace with your actual Supabase URL and Anon key.
+
+4.  **Run the development server:**
+    ```bash
+    npm run dev
+    ```
+    The application will be available at `http://localhost:5173` (or another port if 5173 is busy).
+
+## Building for Production
+
+To create a production build:
+```bash
+npm run build
+```
+The static files will be generated in the `dist` directory.
+
+## Deployment to Cloudflare Pages
+
+This project is ready to be deployed to Cloudflare Pages.
+
+1.  **Push your code** to a GitHub or GitLab repository.
+2.  In the Cloudflare dashboard, go to **Workers & Pages** > **Create application** > **Pages** > **Connect to Git**.
+3.  **Select your repository.**
+4.  **Configure build settings:**
+    *   **Production branch:** Choose the branch you want to deploy (e.g., `main`, `jules`).
+    *   **Framework preset:** Select `Vite`. Cloudflare should auto-fill the rest.
+    *   **Build command:** Verify it's `npm run build` or `vite build`.
+    *   **Build output directory:** Verify it's `dist`.
+    *   **Environment Variables:**
+        *   Add the following environment variables in the Cloudflare Pages settings for your project (Settings > Environment variables > Production and Preview):
+            *   `VITE_SUPABASE_URL`: Your Supabase project URL (e.g., `https://mjwbhlyrfoglbjiwsvqy.supabase.co`)
+            *   `VITE_SUPABASE_ANON_KEY`: Your Supabase anon public key.
+5.  **Save and Deploy.**
+
+### Important for Supabase Authentication:
+*   In your Supabase project dashboard (Authentication > URL Configuration), make sure to add your Cloudflare Pages URL (e.g., `https://your-project.pages.dev`) to the "Site URL" and "Additional Redirect URLs" fields. This is crucial for authentication to work correctly after deployment.
+*   Enable any social providers (Google, GitHub, etc.) you wish to use under Supabase > Authentication > Providers, and configure their client IDs and secrets. Add their respective redirect URIs to your Supabase settings and your provider's settings.
